requirements.txt
chromadb==0.4.1
python-dotenv=1.0.0
Flask==3.0.0
gpt4all==2.0.2
langchain==0.0.332
langchainhub==0.1.13
pypdf==3.17.0
